+++ R/register_config_file.R
@@ -203,13 +203,20 @@ get_list_paths <- function(config_list) {
 #' @param df_files A data.frame with 'type' and 'path' columns
 #' or a csv file path as issued from `[check_not_registered_files()]`
 #' or nothing (and it will take the csv file in "dev/")
-#' @param flat_file_path Character. Usually set to `"keep"` for users. You can use the name of the origin flat file but this is more of an internal use, as inflating the flat file should have the same result.
+#' @param flat_file_path Character. Usually set to `"keep"` for users.
+#' You can use the name of the origin flat file but this is more of
+#' an internal use, as inflating the flat file should have the same result.
 #' @param state Character. Whether if the flat file is `active` or `deprecated`.
-#' @param force Logical. Whether to force writing the configuration file even is some files do not exist.
-#' @param clean Logical. Delete list associated to a specific flat file before updating the whole list.
-#' Default is set to TRUE during `inflate()` of a specific flat file, as the list should only contain files created during the inflate.
-#' This could be set to FALSE with a direct use of `df_to_config()` too. This is forced to FALSE for "keep" section.
-#' @param inflate_parameters list of parameters passed through a call to `inflate()`
+#' @param force Logical. Whether to force writing the configuration
+#' file even is some files do not exist.
+#' @param clean Logical (TRUE, FALSE) or character ("ask", "yes", "no).
+#' Whether to delete files created in a previous version of this flat file
+#' Default is set to "ask" during `inflate()` of a specific flat file,
+#' as the list should only contain files created during the inflate.
+#' This could be set to FALSE with a direct use of `df_to_config()` too.
+#' This is forced to FALSE for the "keep" section.
+#' @param inflate_parameters list of parameters passed through
+#' a call to `inflate()`
 
 #' @importFrom stats setNames
 #' @importFrom utils read.csv
@@ -219,8 +226,10 @@ get_list_paths <- function(config_list) {
 #' Side effect: create a yaml config file.
 #'
 #' @seealso
-#'   [check_not_registered_files()] for the list of files not already associated with a flat file in the config file,
-#'   [register_all_to_config()] for automatically registering all files already present in the project
+#'   [check_not_registered_files()] for the list of files not already
+#'  associated with a flat file in the config file,
+#'   [register_all_to_config()] for automatically registering
+#'  all files already present in the project
 #'
 #' @noRd
 
@@ -228,9 +237,12 @@ df_to_config <- function(df_files,
                          flat_file_path = "keep",
                          state = c("active", "deprecated"),
                          force = FALSE,
-                         clean = TRUE,
+                         clean = "ask",
                          inflate_parameters = NULL) {
-  config_file <- getOption("fusen.config_file", default = "dev/config_fusen.yaml")
+  config_file <- getOption(
+    "fusen.config_file",
+    default = "dev/config_fusen.yaml"
+  )
   state <- match.arg(state, several.ok = FALSE)
 
   # User entry verifications
@@ -242,7 +254,10 @@ df_to_config <- function(df_files,
   if (!is.data.frame(df_files) && file.exists(df_files)) {
     df_files <- read.csv(df_files, stringsAsFactors = FALSE)
   } else if (!is.data.frame(df_files) && !file.exists(df_files)) {
-    stop("'", df_files, "' does not exist. You can run `check_not_registered_files()` before.")
+    stop(
+      "'", df_files, "' does not exist. You can run ",
+      "`check_not_registered_files()` before."
+    )
   }
 
   # Then if is.data.frame(df_files)
@@ -270,7 +285,11 @@ df_to_config <- function(df_files,
       )
       if (isTRUE(force)) {
         cli_alert_warning(
-          paste(msg, "\nHowever, you forced to write it in the yaml file with `force = TRUE`.")
+          paste(
+            msg,
+            "\nHowever, you forced to write it in the yaml",
+            " file with `force = TRUE`."
+          )
         )
       } else {
         stop(msg)
@@ -287,7 +306,12 @@ df_to_config <- function(df_files,
     )
   }
 
-  if (!all(grepl("^R$|^r$|^test$|^tests$|^vignette$|^vignettes$", df_files[["type"]]))) {
+  if (!all(
+    grepl(
+      "^R$|^r$|^test$|^tests$|^vignette$|^vignettes$",
+      df_files[["type"]]
+    )
+  )) {
     stop("Only types 'R', 'test' or 'vignette' are allowed")
   }
   all_exists <- file.exists(df_files[["path"]])
@@ -306,7 +330,10 @@ df_to_config <- function(df_files,
 
     if (isTRUE(force)) {
       cli_alert_warning(
-        paste(msg, "\nHowever, you forced to write it in the yaml file with `force = TRUE`.")
+        paste(
+          msg,
+          "\nHowever, you forced to write it in the yaml file with `force = TRUE`."
+        )
       )
     } else {
       stop(msg)
@@ -314,7 +341,11 @@ df_to_config <- function(df_files,
   }
 
   if (!is.null(inflate_parameters) & flat_file_path == "keep") {
-    stop("The purpose of using \"keep\" is to store files created without inflate(). Therefore it is not allowed to provide inflate_parameters")
+    stop(
+      "The purpose of using \"keep\" is to store files",
+      " created without inflate(). ",
+      " Therefore it is not allowed to provide inflate_parameters"
+    )
   }
 
   # Remove common part between config_file and all path
@@ -333,30 +364,47 @@ df_to_config <- function(df_files,
   }
 
   # All origin path should exist, if not "keep"
-  if (!isTRUE(force) || isTRUE(all(file.exists(df_files$origin[df_files$origin != "keep"])))) {
+  if (
+    !isTRUE(force) ||
+      isTRUE(all(file.exists(df_files$origin[df_files$origin != "keep"])))
+  ) {
     df_files$origin[df_files$origin != "keep"] <- gsub(
       paste0(normalize_path_winslash("."), "/"),
       "",
-      normalize_path_winslash(df_files$origin[df_files$origin != "keep"], mustWork = TRUE),
+      normalize_path_winslash(
+        df_files$origin[df_files$origin != "keep"],
+        mustWork = TRUE
+      ),
       fixed = TRUE
     )
-  } else if (isFALSE(all(file.exists(df_files$origin[df_files$origin != "keep"])))) {
+  } else if (
+    isFALSE(all(file.exists(df_files$origin[df_files$origin != "keep"])))
+  ) {
     warning(
       "Please open a bug on {fusen} package with this complete message:\n",
       "There is an error in the df_to_config process.\n",
       "Files origin do not exist but will be registered as is in the config file:\n",
-      paste(df_files$origin[!file.exists(df_files$origin[df_files$origin != "keep"])],
+      paste(
+        df_files$origin[
+          !file.exists(df_files$origin[df_files$origin != "keep"])
+        ],
         collapse = ", "
       )
     )
   }
 
   if (any(duplicated(df_files$path))) {
-    msg <- paste("Some paths appear multiple times in df_files. Please remove duplicated rows: ", paste(unique(df_files$path[duplicated(df_files$path)]), collapse = ", "))
+    msg <- paste(
+      "Some paths appear multiple times in df_files. Please remove duplicated rows: ",
+      paste(unique(df_files$path[duplicated(df_files$path)]), collapse = ", ")
+    )
 
     if (isTRUE(force)) {
       cli_alert_warning(
-        paste(msg, "\nHowever, you forced to write it in the yaml file with `force = TRUE`.")
+        paste(
+          msg,
+          "\nHowever, you forced to write it in the yaml file with `force = TRUE`."
+        )
       )
     } else {
       stop(msg)
@@ -376,11 +424,16 @@ df_to_config <- function(df_files,
           yaml_paths[!all_exists],
           collapse = ", "
         ), ".\n",
-        "Please open the configuration file: ", config_file, " to verify, and delete the non-existing files if needed."
+        "Please open the configuration file: ",
+        config_file,
+        " to verify, and delete the non-existing files if needed."
       )
       if (isTRUE(force)) {
         cli_alert_warning(
-          paste(msg, "However, you forced to write it in the yaml file with `force = TRUE`.")
+          paste(
+            msg,
+            "However, you forced to write it in the yaml file with `force = TRUE`."
+          )
         )
       } else {
         stop(msg)
@@ -413,7 +466,7 @@ df_to_config <- function(df_files,
   if (length(all_modified) != 0) {
     cli_alert_info(
       paste0(
-        "Some files group already existed and were ", ifelse(isTRUE(clean), "overwritten: ", "modified: "),
+        "Some files group already existed and were modified: ",
         paste(all_modified, collapse = ", ")
       )
     )
@@ -482,16 +535,18 @@ files_list_to_vector <- function(list_of_files) {
 #' @param complete_yaml The list as output of config_yaml file
 #' @param flat_file_path The group to update
 #' @param state Character. See `df_to_config()`.
-#' @param clean Logical. See `df_to_config()`.
+#' @param clean Logical (TRUE, FALSE) or character ("ask", "yes", "no).
+#'  See `df_to_config()`.
 #' @param inflate_parameters See `df_to_config()`.
 #' @importFrom cli cli_alert_warning cli_alert_success
 #' @noRd
-update_one_group_yaml <- function(df_files,
-                                  complete_yaml,
-                                  flat_file_path,
-                                  state = c("active", "deprecated"),
-                                  clean = TRUE,
-                                  inflate_parameters = NULL) {
+update_one_group_yaml <- function(
+    df_files,
+    complete_yaml,
+    flat_file_path,
+    state = c("active", "deprecated"),
+    clean = "ask",
+    inflate_parameters = NULL) {
   state <- match.arg(state, several.ok = FALSE)
   all_keep_before <- complete_yaml[[basename(flat_file_path)]]
 
@@ -499,7 +554,7 @@ update_one_group_yaml <- function(df_files,
   df_files_filtered <- df_files[df_files[["origin"]] == flat_file_path, ]
 
   # All already in the list will be deleted except if clean is FALSE
-  if (isTRUE(clean)) {
+  if (isTRUE(clean) || clean %in% c("yes", "ask")) {
     this_group_list <- list(
       path = flat_file_path,
       state = state,
@@ -513,7 +568,7 @@ update_one_group_yaml <- function(df_files,
         grepl("^vignette$|^vignettes$", df_files_filtered[["type"]])
       ])
     )
-  } else {
+  } else if (isFALSE(clean) || clean %in% c("no")) {
     this_group_list <- list(
       path = flat_file_path,
       state = state,
@@ -535,51 +590,111 @@ update_one_group_yaml <- function(df_files,
       )),
       vignettes = unique(c(
         # new ones
-        df_files_filtered[["path"]][grepl("^vignette$|^vignettes$", df_files_filtered[["type"]])],
+        df_files_filtered[["path"]][
+          grepl(
+            "^vignette$|^vignettes$",
+            df_files_filtered[["type"]]
+          )
+        ],
         # previous ones
         unlist(all_keep_before[["vignettes"]])
       ))
     )
+  } else {
+    stop("clean should be TRUE, FALSE, 'yes', 'no' or 'ask'")
   }
 
   this_group_list_return <- this_group_list
-  this_group_list_message <- this_group_list
   if (!is.null(inflate_parameters)) {
-    this_group_list_return <- c(this_group_list, list(inflate = inflate_parameters))
-    this_group_list_message <- c(this_group_list, list(inflate = "each parameter"))
+    this_group_list_return <- c(
+      this_group_list, list(inflate = inflate_parameters)
+    )
   }
 
   # Messages only
-  all_names <- names(this_group_list_message)
-  those_removed <- lapply(
+  all_names <- c("R", "tests", "vignettes")
+  # names(this_group_list_message)
+  files_removed <- lapply(
     all_names,
     function(x) {
       setdiff(
         all_keep_before[[x]],
-        this_group_list_message[[x]]
+        this_group_list_return[[x]]
       )
     }
   ) %>%
     setNames(all_names)
+  files_removed_vec <- files_list_to_vector(files_removed)
 
-  those_removed_vec <- files_list_to_vector(those_removed)
-  those_added <- lapply(
+  files_added <- lapply(
     all_names,
     function(x) {
       setdiff(
-        this_group_list_message[[x]],
+        this_group_list_return[[x]],
         all_keep_before[[x]]
       )
     }
   ) %>%
     setNames(all_names)
-  those_added_vec <- files_list_to_vector(those_added)
+  files_added_vec <- files_list_to_vector(files_added)
 
-  if (!is.null(those_removed_vec) || length(those_removed_vec) != 0) {
-    silent <- lapply(paste(those_removed_vec, "was removed from the config file"), cli_alert_warning)
+  if (!is.null(files_removed_vec) || length(files_removed_vec) != 0) {
+    if (clean == "ask") {
+      cli_alert_warning(
+        paste0(
+          "Some files are not anymore created from ",
+          flat_file_path, ".\n",
+          "You may have rename some functions or moved them to another flat:",
+          "\n",
+          paste(files_removed_vec, collapse = ", "), ".\n\n",
+          "Below, you are ask if you want to remove them from the repository.",
+          "\n\n",
+          "Note: to not see this message again, use `clean = TRUE` in the ",
+          "`inflate()` command of this flat file : ", flat_file_path, ".\n",
+          "Use with caution. ",
+          "It is recommended to use git to check the changes...\n"
+        )
+      )
+      sure <- paste(
+        paste(files_removed_vec, collapse = ", "),
+        "\nDo you want to remove all these files from the repository? (y/n)\n"
+      )
+
+      do_it <- readline(sure) == "y" || readline(sure) == "yes"
+    } else if (isTRUE(clean) || clean == "yes") {
+      do_it <- TRUE
+    } else if (isFALSE(clean) || clean == "no") {
+      do_it <- FALSE
+    } else {
+      stop("clean should be TRUE, FALSE, 'yes', 'no' or 'ask'")
+    }
+
+    if (isTRUE(do_it)) {
+      files_removed_filename <- unlist(files_removed)
+      file.remove(files_removed_filename)
+
+      silent <- lapply(
+        paste(
+          files_removed_vec, "was removed from the config file",
+          "and from the repository"
+        ),
+        cli_alert_warning
+      )
+    } else if (isFALSE(do_it)) {
+      silent <- lapply(
+        paste(
+          files_removed_vec, "was removed from the config file",
+          "but kept in the repository"
+        ),
+        cli_alert_warning
+      )
+    }
   }
-  if (!is.null(those_added_vec) || length(those_added_vec) != 0) {
-    silent <- lapply(paste(those_added_vec, "was added to the config file"), cli_alert_success)
+  if (!is.null(files_added_vec) || length(files_added_vec) != 0) {
+    silent <- lapply(
+      paste(files_added_vec, "was added to the config file"),
+      cli_alert_success
+    )
   }
 
   return(this_group_list_return)