+++ NEWS.md
@@ -3,9 +3,11 @@
 ## Breaking changes
 
 - {fusen} now relies on {lightparser} instead of {parsermd} to parse flat file. This allows to avoid installation problems with {parsermd}, which is not updated anymore. As {lightparser} is lighter, this may have unattended effects on specific flat file cases. Please report any issue you may encounter. (#233)
+- `inflate_all*()` does not use parameter `clean` anymore. Use `check_unregistered` instead to check if all files are registered in the configuration file.
 
 ## New features
 
+- `inflate()` detects functions renamed or removed and allow to clean the package repository (#24)
 - Allow `organisation` in `init_share_on_github()` to send to a GitHub organisation
 - Fix `load_flat_functions()` to work with VSCode